% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-geno.R
\docType{data}
\name{geno_f2}
\alias{geno_f2}
\title{Example F2 Genotype Data}
\format{
A data frame with individuals in rows and markers in columns.
}
\source{
Simulated or experimental population.
}
\usage{
geno_f2
}
\description{
A toy dataset of genotypes from an F2 population.
}
\examples{
data(geno_f2)
head(geno_f2)
}
\keyword{datasets}
