% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-geno.R
\docType{data}
\name{geno_f6}
\alias{geno_f6}
\title{Example F6 Genotype Data}
\format{
A data frame with individuals in rows and markers in columns.
}
\source{
Simulated or experimental population.
}
\usage{
geno_f6
}
\description{
A toy dataset of genotypes from an F6 population.
}
\examples{
data(geno_f6)
head(geno_f6)
}
\keyword{datasets}
