% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{placeholder_cpp}
\alias{placeholder_cpp}
\title{A no-op function just to give the package something to compile.}
\usage{
placeholder_cpp()
}
\description{
A no-op function just to give the package something to compile.
}
