% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_methods.R
\name{plot_simulation}
\alias{plot_simulation}
\title{Plot simulated vs. estimated map positions}
\usage{
plot_simulation(df)
}
\arguments{
\item{df}{A data frame with numeric columns `simulated` and `estimated`}
}
\description{
Plot simulated vs. estimated map positions
}
