% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selfmap-package.R
\docType{package}
\name{selfmap-package}
\alias{selfmap}
\alias{selfmap-package}
\title{selfmap: EM-Based Recombination Fraction Estimation in Selfing Populations}
\description{
Provides efficient algorithms for estimating recombination fractions in partially inbred (e.g., F2, F6) populations using hidden Markov models and the expectation-maximization (EM) algorithm. Includes functions for genotype data preprocessing, forward–backward inference in HMMs, and map accuracy visualization. Implements core computations in C++ via Rcpp.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mmollina/selfmap}
  \item Report bugs at \url{https://github.com/mmollina/selfmap/issues}
}

}
\author{
\strong{Maintainer}: Marcelo Mollinari \email{mmollin@ncsu.edu}

}
\keyword{internal}
